apply changes to nest-city-account
