+++ nest-city-account/src/api-jwt-tokens/api-jwt-tokens.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common'
+
+import ApiJwtTokensService from './api-jwt-tokens.service'
+
+@Module({
+  providers: [ApiJwtTokensService],
+  exports: [ApiJwtTokensService],
+})
+export default class ApiJwtTokensModule {}
