+++ nest-city-account/src/nases/nases.module.ts
@@ -1,11 +1,12 @@
 import { Module } from '@nestjs/common'
 
+import ApiJwtTokensModule from '../api-jwt-tokens/api-jwt-tokens.module'
 import ClientsModule from '../clients/clients.module'
 import { PrismaModule } from '../prisma/prisma.module'
 import { NasesService } from './nases.service'
 
 @Module({
-  imports: [PrismaModule, ClientsModule],
+  imports: [PrismaModule, ClientsModule, ApiJwtTokensModule],
   providers: [NasesService],
   exports: [NasesService],
 })
