+++ nest-city-account/src/nases/nases.service.spec.ts
@@ -1,6 +1,7 @@
 import { createMock } from '@golevelup/ts-jest'
 import { Test, TestingModule } from '@nestjs/testing'
 
+import ApiJwtTokensService from '../api-jwt-tokens/api-jwt-tokens.service'
 import ClientsService from '../clients/clients.service'
 import ThrowerErrorGuard from '../utils/guards/errors.guard'
 import { NasesService } from './nases.service'
@@ -15,6 +16,7 @@ describe('NasesService', () => {
         NasesService,
         ThrowerErrorGuard,
         { provide: ClientsService, useValue: createMock<ClientsService>() },
+        { provide: ApiJwtTokensService, useValue: createMock<ApiJwtTokensService>() },
       ],
     }).compile()
